tests/wpt/web-platform-tests/app-history/navigate-event/transitionWhile-on-synthetic-event.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  // We need to grab an AppHistoryDestination to construct the event.
  appHistory.onnavigate = t.step_func_done(e => {
    const event = new AppHistoryNavigateEvent("navigate", {
      destination: e.destination,
      signal: (new AbortController()).signal
    });

    assert_throws_dom("SecurityError", () => event.transitionWhile(Promise.resolve()));
  });
  history.pushState(1, null, "#1");
}, "AppHistoryNavigateEvent's transitionWhile() throws if invoked on a synthetic event");
</script>
